+++ instrumentation/java-streams/build.gradle.kts
@@ -1,0 +1,3 @@
+plugins {
+    `java-library`
+}
