+++ instrumentation/java-streams/src/main/java/io/opentelemetry/instrumentation/hypertrace/java/inputstream/InputStreamUtils.java
@@ -14,7 +14,21 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.instrumentation.hypertrace.apachehttpclient.v4_0;
+package io.opentelemetry.instrumentation.hypertrace.java.inputstream; /*
+                                                                       * Copyright The Hypertrace Authors
+                                                                       *
+                                                                       * Licensed under the Apache License, Version 2.0 (the "License");
+                                                                       * you may not use this file except in compliance with the License.
+                                                                       * You may obtain a copy of the License at
+                                                                       *
+                                                                       *     http://www.apache.org/licenses/LICENSE-2.0
+                                                                       *
+                                                                       * Unless required by applicable law or agreed to in writing, software
+                                                                       * distributed under the License is distributed on an "AS IS" BASIS,
+                                                                       * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+                                                                       * See the License for the specific language governing permissions and
+                                                                       * limitations under the License.
+                                                                       */
 
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.api.common.AttributeKey;
